arreglos buscador v3

+++ app/Http/Controllers/FavoritoController.php
@@ -3,7 +3,7 @@
 namespace App\Http\Controllers;
 
 use App\Http\Controllers\Controller;
-use App\Models\favoritos;
+use App\Models\Favorito;
 use Illuminate\Http\Request;
 
 class FavoritoController extends Controller
@@ -11,7 +11,7 @@ class FavoritoController extends Controller
     // Añadir la receta a favoritos
     public function addFavoritos(Request $request) {
 
-        $favorito=new favoritos();
+        $favorito=new Favorito();
 
         $favorito->id_receta=$request->id_receta;
         $favorito->id_usuario=$request->id_usuario;
@@ -23,9 +23,15 @@ class FavoritoController extends Controller
     // Eliminar la receta de favoritos
     public function removeFavoritos($id_receta, $id_usuario) {
 
-        $favorito=favoritos::where('id_Receta',$id_receta)
-                            ->where('id_Usuario',$id_usuario)
+        $favorito=Favorito::where('id_Receta', $id_receta)
+                            ->where('id_Usuario', $id_usuario)
                             ->first();
+
+        if(!$favorito){ 
+            return response()->json([
+                'message' => 'No existe el favorito'
+            ], 404);
+        }
 
         $favorito->delete();
